+++ Controllers/ShowTicket.php
@@ -69,6 +69,7 @@ class ShowTicket extends Controller
         $saveResult = $this->showTicketService->saveTicket($input['id'], $input['key'], $input['value']);
         return response()->json(['ticket' => $saveResult]);
     }
+
     /**
      * Saves ticket headline.
      * @param string[] $input The input for saving:
@@ -80,6 +81,23 @@ class ShowTicket extends Controller
     {
         $deleteResult = $this->showTicketService->deleteTicket($input['id']);
         return response()->json(['ticket' => $deleteResult]);
+    }
+
+    /**
+     * Saves ticket headline.
+     *
+     * @param string[] $input The input for getting tags:
+     *                     - 'id': The project id.
+     *
+     * @return JsonResponse The JSON response containing the list of tags or an empty array.
+     */
+    public function getTags(array $input): JsonResponse
+    {
+        $tags = [];
+        $ticketTags = $this->ticketRepository->getTags($input['projectId']);
+        $tags = $this->explodeAndMergeTags($ticketTags, $tags);
+        $uniqueTags = array_unique($tags);
+        return response()->json(['tags' => $uniqueTags]);
     }
 
     // phpcs:disable
@@ -130,10 +148,6 @@ class ShowTicket extends Controller
             $ticket = $this->showTicketService->getTicket($_GET['ticketId']);
 
             if ($ticket) {
-                $tags = [];
-                $ticketTags = $this->ticketRepository->getTags($ticket->projectId);
-                $tags = $this->explodeAndMergeTags($ticketTags, $tags);
-                $uniqueTags = array_unique($tags);
                 $statusLabels = $this->showTicketService->getStatusLabels($ticket->projectId);
                 $priorityLabels = $this->showTicketService->getPriorityLabels();
                 $this->template->assign('ticket', $ticket);
@@ -142,7 +156,6 @@ class ShowTicket extends Controller
                 $this->tpl->assign('sprints', $this->sprintService->getAllSprints($ticket->projectId));
                 $this->tpl->assign('allUsers', $this->userService->getAll());
                 $this->tpl->assign('relatedTickets', $this->ticketService->getAllPossibleParents($ticket));
-                $this->tpl->assign('tags', $uniqueTags);
                 $milestones = $this->ticketService->getAllMilestones(['sprint' => '', 'type' => 'milestone', 'currentProject' => $ticket->projectId]);
                 $this->tpl->assign('milestones', $milestones);
                 $this->tpl->assign('files', $this->filesRepo->getFilesByModule('ticket', $ticket->id));